Web/backend/party_currency_backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-allauth>=0.54.0
dj-rest-auth
psycopg2-binary
gunicorn
python-dotenv
whitenoise
PyJWT
requests
requests-oauthlib
cryptography
Pillow
pytz
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
  